Version 1.0beta4
Changed the way we parse the lines: instead of reading the line length value from the BASIC line (because sometimes it is wrong), we read until we hit \r (carriage return, new line marker on Spectrum). This is in line with what LIST and LLIST do, because they have no problems printing programs with such errors.